+++ src/errors.ts
@@ -1,7 +1,7 @@
 /** @format */
 
 export class InvalidPropError extends Error {
-  constructor(typeName, key: string) {
+  constructor(typeName: string, key: string) {
     super(`In property ${key}: ${typeName} is not a primitive type nor a Typegoose schema (Not extending it).`);
   }
 }
