+++ src/test/models/person.ts
@@ -1,62 +1,31 @@
-import * as mongoose from 'mongoose';
+/** @format */
+
 import * as tg from '../../typegoose';
-import { Car } from './car';
+import { PersistentModel } from './PersistentModel';
 
 // add a pre-save hook to PersistentModel
 @tg.pre<PersistentModel>('save', function(next) {
-    if (!this.createdAt) {
-        this.createdAt = new Date();
-    }
-    next();
+  if (!this.createdAt) {
+    this.createdAt = new Date();
+  }
+  next();
 })
-export abstract class PersistentModel extends tg.Typegoose {
-    @tg.prop()
-    createdAt: Date;
-
-    @tg.arrayProp({ itemsRef: Car })
-    cars?: tg.Ref<Car>[];
-
-    // define an 'instanceMethod' that will be overwritten
-    @tg.instanceMethod
-    getClassName() {
-        return 'PersistentModel';
-    }
-
-    // define an 'instanceMethod' that will be overwritten
-    @tg.staticMethod
-    static getStaticName() {
-        return 'PersistentModel';
-    }
-
-    // define an instanceMethod that is called by the derived class
-    @tg.instanceMethod
-    addCar(this: tg.InstanceType<PersistentModel>, car: Car) {
-        if (!this.cars) {
-            this.cars = [];
-        }
-
-        this.cars.push(car);
-
-        return this.save();
-    }
-}
-
 export class Person extends PersistentModel {
-    // add new property
-    @tg.prop({ required: true, validate: /^[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,}$/ })
-    email: string;
+  // add new property
+  @tg.prop({ required: true, validate: /^[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,}$/ })
+  email: string;
 
-    // override instanceMethod
-    @tg.instanceMethod
-    getClassName() {
-        return 'Person';
-    }
+  // override instanceMethod
+  @tg.instanceMethod
+  getClassName() {
+    return 'Person';
+  }
 
-    // override staticMethod
-    @tg.staticMethod
-    static getStaticName() {
-        return 'Person';
-    }
+  // override staticMethod
+  @tg.staticMethod
+  static getStaticName() {
+    return 'Person';
+  }
 }
 
 export const model = new Person().getModelForClass(Person);
